+++ deps/libiconv/lib/cp1258.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2004 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2004, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,8 +14,7 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
- * Fifth Floor, Boston, MA 02110-1301, USA.
+ * If not, see <http://www.gnu.org/licenses/>.
  */
 
 /*
@@ -72,7 +71,7 @@ static const unsigned short cp1258_2uni[128] = {
    character, or 0 if none. */
 
 static int
-cp1258_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp1258_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   unsigned short wc;
@@ -201,7 +200,7 @@ static const unsigned char cp1258_page20[48] = {
 };
 
 static int
-cp1258_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp1258_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   unsigned char c = 0;
   if (wc < 0x0080) {
