+++ deps/libiconv/lib/cp860.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2002 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2002, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,8 +14,7 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
- * Fifth Floor, Boston, MA 02110-1301, USA.
+ * If not, see <http://www.gnu.org/licenses/>.
  */
 
 /*
@@ -50,7 +49,7 @@ static const unsigned short cp860_2uni[128] = {
 };
 
 static int
-cp860_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp860_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   if (c < 0x80)
@@ -120,7 +119,7 @@ static const unsigned char cp860_page25[168] = {
 };
 
 static int
-cp860_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp860_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   unsigned char c = 0;
   if (wc < 0x0080) {
