+++ deps/libiconv/lib/cp949.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2005, 2007 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2005, 2007, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,8 +14,7 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
- * Fifth Floor, Boston, MA 02110-1301, USA.
+ * If not, see <http://www.gnu.org/licenses/>.
  */
 
 /*
@@ -38,7 +37,7 @@
 #include "uhc_2.h"
 
 static int
-cp949_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp949_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   /* Code set 0 (ASCII) */
@@ -79,7 +78,7 @@ cp949_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 }
 
 static int
-cp949_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp949_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   unsigned char buf[2];
   int ret;
